+++ web/app/scripts/compiled-es5/directives/proportion-field.js
@@ -22,7 +22,7 @@ _module.directive('proportionField', [function () {
         templateUrl: 'templates/incuna-surveys/form/proportion-field.html',
         link: function link(scope) {
             scope.fields = [];
-            scope.$watch('options', function (options, oldValue) {
+            scope.$watch('options', function (options) {
                 if (options.fieldOptions) {
                     scope.title = options.fieldOptions.label;
                 }
@@ -30,14 +30,13 @@ _module.directive('proportionField', [function () {
                     options.choices.forEach(function (choice, index) {
                         scope.fields[index] = Object.assign({}, options.fieldOptions, {
                             label: choice,
-                            id: options.autoId + '-' + index,
-                            hex: Math.floor(Math.random(index) * 16777215).toString(16)
+                            id: options.autoId + '-' + index
                         });
                     });
                 }
             });
 
-            scope.$watch('options.fieldOptions.errors', function (errors, oldValue) {
+            scope.$watch('options.fieldOptions.errors', function (errors) {
                 if (errors) {
                     scope.fields.forEach(function (options, index) {
                         options.errors = errors[index];
@@ -45,7 +44,7 @@ _module.directive('proportionField', [function () {
                 }
             });
 
-            scope.$watch('model', function (values, oldValue) {
+            scope.$watch('model', function (values) {
                 scope.total = Object.keys(values).reduce(function (value, key) {
                     return value + (values[key] ? parseInt(values[key], 10) : 0);
                 }, 0);