+++ web/app/scripts/src/directives/proportion-field.es.js
@@ -17,26 +17,25 @@ module.directive('proportionField', [
             templateUrl: 'templates/incuna-surveys/form/proportion-field.html',
             link: function (scope) {
                 scope.fields = [];
-                scope.$watch('options', (options, oldValue) => {
+                scope.$watch('options', (options) => {
                     if (options.fieldOptions) {
                         scope.title = options.fieldOptions.label;
                     }
-                    if (options.choices) { 
+                    if (options.choices) {
                         options.choices.forEach((choice, index) => {
                             scope.fields[index] = Object.assign(
                                 {},
                                 options.fieldOptions,
                                 {
                                     label: choice,
-                                    id: `${options.autoId}-${index}`,
-                                    hex: Math.floor(Math.random(index)*16777215).toString(16)
+                                    id: `${options.autoId}-${index}`
                                 }
                             );
                         });
                     }
                 });
 
-                scope.$watch('options.fieldOptions.errors', (errors, oldValue) => {
+                scope.$watch('options.fieldOptions.errors', (errors) => {
                     if (errors) {
                         scope.fields.forEach((options, index) => {
                             options.errors = errors[index];
@@ -44,7 +43,7 @@ module.directive('proportionField', [
                     }
                 });
 
-                scope.$watch('model', (values, oldValue) => {
+                scope.$watch('model', (values) => {
                     scope.total = Object.keys(values).reduce(
                         ( value, key ) => value + (values[key] ? parseInt(values[key], 10) : 0),
                         0
@@ -56,7 +55,6 @@ module.directive('proportionField', [
                         })
                     }
                 }, true);
-                
             }
         };
     }